+++ quilt-sync/src-tauri/src/model.rs
@@ -103,6 +103,14 @@ pub trait QuiltModel {
         Ok(package.lineage().await?)
     }
 
+    async fn get_installed_package_current_revision(
+        &self,
+        package: &quilt::InstalledPackage,
+        lineage: &quilt::lineage::PackageLineage,
+    ) -> Result<Option<quilt::flow::Revision>, Error> {
+        Ok(package.current_revision(lineage).await?)
+    }
+
     async fn get_installed_package_records(
         &self,
         package: &quilt::InstalledPackage,